Move types to `types`
